Refactor acceptance specs for Personal details
- we would like the "submits a valid application" spec to grow into a
complete end-to-end spec as the application gets bigger
- we like having rich before: blocks and skinny asserting tests
underneath them
- we like making a user the subject of the spec, eg "A candidate
completing a form" is better than "A candidate completes the form"
